refactor(cli): core.py helpers + registrar loop foundation (#495)
Phase 0 of the #495 CLI-monolith extraction. Pure relocation, no CLI
behavior change.

- New agentwire/core.py holds ~32 stateless helpers + AgentCommand,
  CONFIG_DIR, _UNATTENDED_ENV_KEYS (tmux probes, env/agent-command,
  config/paths, session metadata, session/machine resolution, output/
  format utils). __main__ re-imports every moved name.
- prompt_router.py and diff_cli.py consume core directly instead of
  reimplementing metadata reads / importing from __main__.
- build_parser() now drives diff/prompts/msg/limits through a _REGISTRARS
  loop; limits_cli gains register_limits_parser() mirroring the others.
- Tests that monkeypatch moved helpers repointed to agentwire.core.

Deliverable #3 (swap inline tmux has-session strings to
tmux_session_exists) dropped: all 6 are remote-over-SSH / _run_remote,
not local-helper bypasses; #495 premise (b) was incorrect.

Built by dotdev.dev
